add one simple receiver to check the notification
