--- application/template/NGSILD/goTemplate/config.json
@@ -1,11 +1,0 @@
-[
-{
-	"command":"CONNECT_BROKER",
-	"brokerURL": "http://180.179.214.202:8070" 
-},
-{
-	"command":"SET_INPUTS",
-	"type": "test",
-	"id":   "urn:ngsi-ld:Vehicle:A14"
-}
-]
