--- application/template/NGSILD/goTemplate/dockerfile
@@ -1,5 +1,0 @@
-FROM scratch
-
-ADD goTemplate  /
-CMD ["./goTemplate"]
-
