--- test/subscriber/config.json
@@ -1,9 +1,0 @@
-{
-	"my_ip": "127.0.0.1",	
-	"update_broker_url": "http://127.0.0.1:8080/ngsi10",	
-	"subscribe_broker_url": "http://127.0.0.1:8080/ngsi10",	    
-	"logging":{
-		"info":"stdout"
-	}			
-}
-
